# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/auth/sign-in": {
            "post": {
                "description": "Log in to the admin panel, get a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.SingUpResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/links/to": {
            "get": {
                "description": "Link redirect ( for stat )",
                "tags": [
                    "Links"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "the link to which the redirect will be made",
                        "name": "link",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "the page from which the link was clicked",
                        "name": "page",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "description": "Get list Projects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/projects.ProjectsResp"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "description": "Get information about the project by its ID and language, if specified",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Projects"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project language (English by default)",
                        "name": "lang",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/projects.ProjectResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/api/stat/clicks": {
            "get": {
                "description": "Get how many times clicked on links",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/stat/projects": {
            "get": {
                "description": "Get project visits by day in 30 days",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/stat.ProjectsStats"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/stat/update/projects/{uuid}": {
            "put": {
                "description": "Increments the number of project visits by 1 for the current day",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "project id",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful statistics update"
                    }
                }
            }
        },
        "/api/stat/update/visit": {
            "put": {
                "description": "add a visit to the site, gets the IP, country, time of entry and uniqueness of the visit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Unique visit or not (0 - not unique, 1 - unique)",
                        "name": "un",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UTM tag ID (optional parameter)",
                        "name": "utm",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful registration of the session and return of the session ID",
                        "schema": {
                            "$ref": "#/definitions/stat.SessionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/stat/updatevisit/extend": {
            "put": {
                "description": "Extends the time spent on the site by session ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "session",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/api/stat/visits": {
            "get": {
                "description": "Get information about site visits, including the number of visits, unique visits, top countries, top browsers and top OS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stat"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/stat.VisitsResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "auth.SingUpResp": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "token for authorization",
                    "type": "string"
                }
            }
        },
        "projects.ProjectResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "description": "Project's card",
                    "type": "string"
                },
                "categories": {
                    "description": "Category ids",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "description": {
                    "description": "Project's description",
                    "type": "string"
                },
                "features": {
                    "description": "Project's features",
                    "type": "string"
                },
                "id": {
                    "description": "Project's id",
                    "type": "string"
                },
                "photos": {
                    "description": "Project's photos",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "subtitle": {
                    "description": "Project's subtitle",
                    "type": "string"
                },
                "title": {
                    "description": "Project's title",
                    "type": "string"
                },
                "url": {
                    "description": "url preview",
                    "type": "string"
                }
            }
        },
        "projects.ProjectsResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/projects.ProjectsResponse"
                    }
                }
            }
        },
        "projects.ProjectsResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "description": "Category ids",
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "id": {
                    "description": "Project's id",
                    "type": "string"
                },
                "subtitle": {
                    "description": "Project's subtitle",
                    "type": "string"
                },
                "title": {
                    "description": "Project's title",
                    "type": "string"
                },
                "url": {
                    "description": "Url preview",
                    "type": "string"
                }
            }
        },
        "stat.ProjectsStats": {
            "type": "object",
            "properties": {
                "count": {
                    "description": "Count visits",
                    "type": "integer"
                },
                "date": {
                    "description": "Date",
                    "type": "string"
                },
                "uuid": {
                    "description": "Project Id",
                    "type": "string"
                }
            }
        },
        "stat.SessionResponse": {
            "type": "object",
            "properties": {
                "session": {
                    "type": "string"
                }
            }
        },
        "stat.SiteStats": {
            "type": "object",
            "properties": {
                "avg_time_on_site": {
                    "description": "Average time spent on the site",
                    "type": "integer"
                },
                "top_browsers": {
                    "description": "Top browsers",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "top_countries": {
                    "description": "Top countries",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "top_os": {
                    "description": "Top Operating systems",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "total_visits": {
                    "description": "Visits",
                    "type": "integer"
                },
                "total_visits_by_day": {
                    "description": "Visits per day",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "unique_visits": {
                    "description": "Unique visits",
                    "type": "integer"
                },
                "unique_visits_by_day": {
                    "description": "Unique visits per day",
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "visits_details_by_days": {
                    "description": "Full statistics of visits by users",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/stat.VisitsDetails"
                    }
                }
            }
        },
        "stat.Visits": {
            "type": "object",
            "properties": {
                "browser": {
                    "description": "Browser (Chrome, Firefox, etc.)",
                    "type": "string"
                },
                "country": {
                    "description": "Short country code (EE, DE, etc.)",
                    "type": "string"
                },
                "ip": {
                    "description": "The IP address from which",
                    "type": "string"
                },
                "os": {
                    "description": "Operating system (Windows, macOS, etc.)",
                    "type": "string"
                },
                "platform": {
                    "description": "Platform (Linux, Macintosh, iPhone)",
                    "type": "string"
                },
                "time_entry": {
                    "description": "The time at which the person entered",
                    "type": "string"
                },
                "time_leaving": {
                    "description": "The time when the user left the site",
                    "type": "string"
                },
                "uid": {
                    "description": "Unique identifier",
                    "type": "string"
                },
                "unique": {
                    "description": "Has the user already logged in or not",
                    "type": "boolean"
                },
                "utm": {
                    "description": "the UTM tag id came in",
                    "type": "string"
                }
            }
        },
        "stat.VisitsDetails": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "Date",
                    "type": "string"
                },
                "details": {
                    "description": "Details",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/stat.Visits"
                    }
                }
            }
        },
        "stat.VisitsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/stat.SiteStats"
                }
            }
        }
    }
}